use 'Unix Makefiles' cmake generator for libyuv build on Android
for some unknown reason cmake unable to find ninja when tries to
build libyuv, regardless it is able to find it in any other cases.
looks like the issue is specific to ExternalProject_Add() call,
and ninja boundled with cmake (installed through AndroidStudio)
cann't be found for such projects (some paths are not propogated).
having ninja in any of system-wide paths solves this issue, but
this is not an option - not every developer has ninja installed.